Added User Choice to Join A Company
Added the view and controller after a user registers and they choose to
not create a company. They are given the choice to join a company and if
they do they are redirected to the choose company page and if they
decline they are sent home because they cant do anything.